Remove secrets store service
