+++ pkg/server/wire.go
@@ -50,7 +50,6 @@ import (
 	"github.com/grafana/grafana/pkg/services/schemaloader"
 	"github.com/grafana/grafana/pkg/services/search"
 	"github.com/grafana/grafana/pkg/services/secrets"
-	"github.com/grafana/grafana/pkg/services/secrets/database"
 	"github.com/grafana/grafana/pkg/services/shorturls"
 	"github.com/grafana/grafana/pkg/services/sqlstore"
 	"github.com/grafana/grafana/pkg/setting"
@@ -145,8 +144,6 @@ var wireBasicSet = wire.NewSet(
 	prometheus.ProvideService,
 	elasticsearch.ProvideService,
 	secrets.ProvideSecretsService,
-	database.ProvideSecretsStore,
-	wire.Bind(new(secrets.SecretsStore), new(*database.SecretsStoreImpl)),
 	grafanads.ProvideService,
 	dashboardsnapshots.ProvideService,
 )
